requirements.txt
pyserial>=3.5
influxdb-client>=1.36.0
PyYAML>=6.0